requirements.txt
django-photologue
